+++ source/runtime/interpret.cpp
@@ -131,7 +131,7 @@ void initMist() {
 }
 #endif
 
-bool Scratch::startScratchProject() {
+void loadUsernameFromSettings() {
     customUsername = "Player";
     useCustomUsername = false;
 
@@ -159,6 +159,10 @@ bool Scratch::startScratchProject() {
             else customUsername = "Player";
         }
     }
+}
+
+bool Scratch::startScratchProject() {
+    loadUsernameFromSettings();
 #ifdef ENABLE_CLOUDVARS
     if (cloudProject && !projectJSON.empty()) initMist();
 #endif
